move VerifyEmail to components/
