+++ pages/strom/verify-email/[verificationKey].tsx
@@ -1,7 +1,7 @@
 import {NextPage} from 'next'
 
 import {PageLayout} from '../../../src/components/PageLayout/PageLayout'
-import {VerifyEmail} from '../../../src/pages/VerifyEmail/VerifyEmail'
+import {VerifyEmail} from '../../../src/components/VerifyEmail/VerifyEmail'
 
 const Verify: NextPage = () => (
   <PageLayout seminarId={1}>
